Azure_CLI_MCP/requirements.txt
mcp
azure-identity
openai
logging
flask
flask-cors
anthropic
python-dotenv